requirements.txt
aiogram==3.17.0
aiosqlite==0.20.0
apscheduler==3.10.4
python-dotenv==1.0.1
uvloop==0.19.0; sys_platform == 'linux' or sys_platform == 'darwin'
